Rename `authenticator` module
