+++ dbaccess/__init__.py
@@ -1,3 +1,2 @@
-from dbaccess.specs import DBSpecs, AWSRDSSpecs
-from dbaccess.authenticator import get_authenticator, AWSRDSAccessAuthenticator
+from dbaccess.authenticate import get_authenticator, AWSRDSAccessAuthenticator
 from dbaccess.access import create_engine_from_specs
